feat: rota para listar todas ações disponiveis

+++ src/InvestimentManager.sql
@@ -16,7 +16,7 @@ CREATE TABLE ativos_disponiveis (
     id INT NOT NULL auto_increment,
     nome VARCHAR(30) NOT NULL,
     qtdeAtivo INT NOT NULL,
-    valor FLOAT NOT NULL,
+    valor DOUBLE NOT NULL,
     PRIMARY KEY(id)
 ) ENGINE=INNODB;
 
@@ -53,7 +53,7 @@ CREATE TABLE investimentos (
     CodCliente INT NOT NULL,
     CodAtivo INT NOT NULL,
 	QtdeAtivo INT NOT NULL,
-    Valor FLOAT NOT NULL,
+    Valor DOUBLE NOT NULL,
     PRIMARY KEY(id),
     FOREIGN KEY (CodCliente)
         REFERENCES clientes (codCliente)
@@ -66,7 +66,7 @@ CREATE TABLE investimentos (
 CREATE TABLE hist_deposito (
     id INT NOT NULL auto_increment,
     CodCliente INT NOT NULL,
-    Valor FLOAT NOT NULL,
+    Valor DOUBLE NOT NULL,
     PRIMARY KEY(id),
     FOREIGN KEY (CodCliente)
         REFERENCES clientes (codCliente)
@@ -76,7 +76,7 @@ CREATE TABLE hist_deposito (
 CREATE TABLE hist_saque (
     id INT NOT NULL auto_increment,
     CodCliente INT NOT NULL,
-    Valor FLOAT NOT NULL,
+    Valor DOUBLE NOT NULL,
     PRIMARY KEY(id),
     FOREIGN KEY (CodCliente)
         REFERENCES clientes (codCliente)
